fix: access classes from internal fabric
